rename worker resources
