+++ report/src/detector-wrapper/module.js
