+++ report/src/detector-wrapper/util.js
